[add] splitting client and pipelinedClient across go files
